jobify-scheduler/pyproject.toml
[tool.poetry]
name = "jobify-scheduler"
version = "0.1.0"
description = ""
authors = ["btkcodedev <btk.codedev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
firebase-admin = "^6.5.0"
python-telegram-bot = "^21.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
google-cloud-firestore = "^2.16.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
